Improve login testing. Rename AuthenticationServiceTest to UserLoginTest in order to allow for more customized authentication service tests. Implement TokenPairValidator to easily test token pairs. (#9)
